Order placed

+++ controller/orderController.js
@@ -5,26 +5,45 @@ import { Cart } from "../models/cart.model.js";
 
 export const getAllOrders = async (req, res) => {
   try {
-    const page = req.query.page || 1;
-    const limit = req.query.limit || 10;
-
+    let { page = 1, limit = 2, search = null } = req.query;
     const skip = (page - 1) * limit;
+    let query = {};
 
-    const orders = await Order.find().skip(skip).limit(limit);
-    const totalOrders = await Order.countDocuments();
-
-    const totalPages = Math.ceil(totalOrders / limit);
-    if (!orders || !orders.length === 0) {
-      return res
-        .status(404)
-        .json({ success: false, message: "orders not available" });
+    if (search) {
+      const searchQuery = new RegExp(search.toString(), "i") || null;
+      query.$or = [
+        { shippingAddress: searchQuery },
+        { billingAddress: searchQuery },
+      ];
     }
 
-    res
-      .status(200)
-      .json({ success: true, message: "Here is the orders", orders,totalOrders:totalOrders,totalPages:totalPages});
+    const [orders, totalOrders] = await Promise.all([
+      Order.find(query).skip(skip).limit(limit),
+      Order.countDocuments(query),
+    ]);
+
+    if (!orders || orders.length === 0) {
+      return res.status(404).json({
+        success: false,
+        message: "orders not available",
+      });
+    }
+
+    const totalPages = Math.ceil(totalOrders / limit);
+    res.status(200).json({
+      success: true,
+      message: "Here is the orders",
+      orders,
+      totalOrders: totalOrders,
+      totalPages: totalPages,
+    });
   } catch (err) {
-    return res.status(500).json({ success: false, message: "Internal issue" });
+    console.log("Error", err);
+    return res.status(500).json({
+      success: false,
+      message: "Internal issue",
+      error: err,
+    });
   }
 };
 
@@ -32,12 +51,15 @@ export const getOrder = async (req, res) => {
   try {
     const { orderId } = req.params;
     if (!orderId) {
-      return res
-        .status(400)
-        .json({ success: false, message: "Please provide order details" });
+      return res.status(400).json({
+        success: false,
+        message: "Please provide order details",
+      });
     }
-    const order = await Order.findOne({ _id: orderId });
-    const orderItems = await OrderItem.find({ orderId });
+    const [order,orderItems] = await Promise.all([
+      Order.findOne({ _id: orderId }),
+      OrderItem.find({ orderId })
+    ])
 
     if (!order) {
       return res
@@ -45,9 +67,12 @@ export const getOrder = async (req, res) => {
         .json({ success: false, message: "Order not available in db" });
     }
 
-    res
-      .status(200)
-      .json({ success: true, message: "Order details", order, orderItems });
+    res.status(200).json({
+      success: true,
+      message: "Order details",
+      order,
+      orderItems,
+    });
   } catch (err) {
     return res.status(500).json({ success: false, message: "Internal issue" });
   }
@@ -60,29 +85,31 @@ export const placeOrder = async (req, res) => {
     const userCart = await Cart.findOne({ userId: user._id });
 
     if (!userCart) {
-      return res
-        .status(400)
-        .json({ success: false, message: "Cart not available" });
+      return res.status(400).json({
+        success: false,
+        message: "Cart not available",
+      });
     }
-
     const cartItems = await CartItem.find({ cartId: userCart._id });
-    // Create a new order
 
     if (!cartItems || cartItems.length === 0) {
-      return res
-        .status(400)
-        .json({ success: false, message: "Cart items not available" });
+      return res.status(400).json({
+        success: false,
+        message: "Cart items not available",
+      });
     }
 
-    const totalAmount = cartItems.reduce(
-      (total, item) => total + item.subTotal + item.tax,
-      0
+    const { totalAmount, totalQty, totalTax } = cartItems.reduce(
+      (acc, item) => {
+        acc.totalAmount += item.subTotal + item.tax;
+        acc.totalQty += item.qty + item.tax;
+        acc.totalTax += item.qty + item.tax;
+
+        return acc;
+      },
+      { totalAmount: 0, totalQty: 0, totalTax: 0 }
     );
-    const totalQty = cartItems.reduce(
-      (total, item) => total + item.qty + item.tax,
-      0
-    );
-    const totalTax = cartItems.reduce((total, item) => total + item.tax, 0);
+
     const newOrder = await Order.create({
       userId: user._id,
       totalAmount: totalAmount,
@@ -93,7 +120,13 @@ export const placeOrder = async (req, res) => {
       billingAddress: user.billingAddress,
     });
 
-    console.log(newOrder);
+    if (!newOrder) {
+      return res.status(500).json({
+        status: false,
+        message: "Error while creating the order",
+      });
+    }
+
     const orderItems = cartItems.map((cartItem) => ({
       orderId: newOrder._id,
       productId: cartItem.productId,
@@ -104,12 +137,19 @@ export const placeOrder = async (req, res) => {
       subTotal: cartItem.subTotal,
     }));
 
-    console.log(orderItems);
+    const placeItems = await OrderItem.insertMany(orderItems);
 
-    await OrderItem.insertMany(orderItems);
+    if (!placeItems || placeItems?.length == 0) {
+      return res.status(500).json({
+        status: false,
+        message: "Error while creating order items",
+      });
+    }
 
-    await Cart.findOneAndDelete({ _id: userCart._id });
-    await CartItem.deleteMany({ _id: userCart._id });
+    await Promise.all([
+      Cart.findOneAndDelete({ _id: userCart._id }),
+      CartItem.deleteMany({ _id: userCart._id }),
+    ]);
 
     return res.status(200).json({
       success: true,
